módulo autenticación ok

+++ views/templates/borrar_consulta.php
@@ -3,6 +3,8 @@ require_once "./controllers/consultaController.php";
 
 $id = $vista[1];
 
-eliminarConsultaPorId($id);
+if(isset($_SESSION["rol"]) and $_SESSION["rol"] == "Admin"){
+  eliminarConsultaPorId($id);
+}
 
 header("Location: $URL/listado_consultas");
